+++ src/PomClass/PreAdHome_PomClass.java
@@ -12,12 +12,7 @@ public class PreAdHome_PomClass
     @FindBy(xpath = "//span[text()='Preadmission']")
     private  WebElement Preadmission;
     
-    @FindBy(xpath = "//ul[@class ='sidebar-menu styledisplay']/li[4]/ul/li[4]/a")
-    private WebElement Applicationform;
-    
-    @FindBy(xpath = "//ul[@class ='sidebar-menu styledisplay']/li[4]/ul/li[4]/ul")
-    private WebElement applicationformsub;
-    
+      
     public PreAdHome_PomClass(WebDriver driver)
     {
     	PageFactory.initElements(driver, this);
@@ -28,15 +23,6 @@ public class PreAdHome_PomClass
 		return Preadmission;
 	}
 
-	public WebElement getApplicationform()
-	{
-		return Applicationform;
-	}
-
-	public WebElement getApplicationformsub()
-	{
-		return applicationformsub;
-	}
