+++ test-output/Suite/Test.html
@@ -55,11 +55,11 @@ function toggleAllBoxes() {
 <body>
 <h2 align='center'>Test</h2><table border='1' align="center">
 <tr>
-<td>Tests passed/Failed/Skipped:</td><td>2/0/0</td>
+<td>Tests passed/Failed/Skipped:</td><td>3/0/0</td>
 </tr><tr>
-<td>Started on:</td><td>Thu Sep 14 09:50:10 IST 2023</td>
+<td>Started on:</td><td>Wed Sep 20 09:38:52 IST 2023</td>
 </tr>
-<tr><td>Total time:</td><td>25 seconds (25129 ms)</td>
+<tr><td>Total time:</td><td>33 seconds (33691 ms)</td>
 </tr><tr>
 <td>Included groups:</td><td></td>
 </tr><tr>
@@ -79,12 +79,17 @@ function toggleAllBoxes() {
 <td title='TestScript.LoginPage.Testapplication()'><b>Testapplication</b><br>Test class: TestScript.LoginPage</td>
 <td></td>
 <td>0</td>
-<td>TestScript.LoginPage@72e5a8e</td></tr>
+<td>TestScript.LoginPage@54e1c68b</td></tr>
 <tr>
 <td title='TestScript.PreAdm_HomePage.Test()'><b>Test</b><br>Test class: TestScript.PreAdm_HomePage</td>
 <td></td>
-<td>11</td>
-<td>TestScript.PreAdm_HomePage@771a660</td></tr>
+<td>10</td>
+<td>TestScript.PreAdm_HomePage@2d52216b</td></tr>
+<tr>
+<td title='TestScript.Application_form.Test()'><b>Test</b><br>Test class: TestScript.Application_form</td>
+<td></td>
+<td>10</td>
+<td>TestScript.Application_form@242b836</td></tr>
 </table><p>
 </body>
 </html>
