+++ test-output/Suite/Test.xml
@@ -1,8 +1,12 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!-- Generated by org.testng.reporters.JUnitXMLReporter -->
-<testsuite ignored="1" hostname="DESKTOP-9JGPINH" failures="0" tests="2" name="Test" time="25.129" errors="0" timestamp="2023-09-14T09:50:36 IST">
-  <testcase classname="TestScript.LoginPage" name="Testapplication" time="0.361"/>
-  <testcase classname="TestScript.PreAdm_HomePage" name="Test" time="11.086"/>
+<testsuite ignored="2" hostname="DESKTOP-9JGPINH" failures="0" tests="3" name="Test" time="33.691" errors="0" timestamp="2023-09-20T09:39:26 IST">
+  <testcase classname="TestScript.LoginPage" name="Testapplication" time="0.38"/>
+  <testcase classname="TestScript.PreAdm_HomePage" name="Test" time="10.786"/>
+  <testcase classname="TestScript.Application_form" name="Test" time="10.588"/>
+  <testcase classname="Generic.BaseClass" name="@BeforeTest Openapplication" time="0.0">
+    <ignored/>
+  </testcase> <!-- @BeforeTest Openapplication -->
   <testcase classname="Generic.BaseClass" name="@BeforeTest Openapplication" time="0.0">
     <ignored/>
   </testcase> <!-- @BeforeTest Openapplication -->
