+++ test-output/emailable-report.html
@@ -9,11 +9,12 @@
 <table>
 <tr><th>Test</th><th># Passed</th><th># Skipped</th><th># Retried</th><th># Failed</th><th>Time (ms)</th><th>Included Groups</th><th>Excluded Groups</th></tr>
 <tr><th colspan="8">Suite</th></tr>
-<tr><td><a href="#t0">Test</a></td><td class="num">2</td><td class="num">0</td><td class="num">0</td><td class="num">0</td><td class="num">25,129</td><td></td><td></td></tr>
+<tr><td><a href="#t0">Test</a></td><td class="num">3</td><td class="num">0</td><td class="num">0</td><td class="num">0</td><td class="num">33,691</td><td></td><td></td></tr>
 </table>
-<table id='summary'><thead><tr><th>Class</th><th>Method</th><th>Start</th><th>Time (ms)</th></tr></thead><tbody><tr><th colspan="4">Suite</th></tr></tbody><tbody id="t0"><tr><th colspan="4">Test &#8212; passed</th></tr><tr class="passedeven"><td rowspan="1">TestScript.LoginPage</td><td><a href="#m0">Testapplication</a></td><td rowspan="1">1694665224606</td><td rowspan="1">361</td></tr><tr class="passedodd"><td rowspan="1">TestScript.PreAdm_HomePage</td><td><a href="#m1">Test</a></td><td rowspan="1">1694665224981</td><td rowspan="1">11086</td></tr></tbody>
+<table id='summary'><thead><tr><th>Class</th><th>Method</th><th>Start</th><th>Time (ms)</th></tr></thead><tbody><tr><th colspan="4">Suite</th></tr></tbody><tbody id="t0"><tr><th colspan="4">Test &#8212; passed</th></tr><tr class="passedeven"><td rowspan="1">TestScript.Application_form</td><td><a href="#m0">Test</a></td><td rowspan="1">1695182955831</td><td rowspan="1">10588</td></tr><tr class="passedodd"><td rowspan="1">TestScript.LoginPage</td><td><a href="#m1">Testapplication</a></td><td rowspan="1">1695182944652</td><td rowspan="1">380</td></tr><tr class="passedeven"><td rowspan="1">TestScript.PreAdm_HomePage</td><td><a href="#m2">Test</a></td><td rowspan="1">1695182945039</td><td rowspan="1">10786</td></tr></tbody>
 </table>
-<h2>Test</h2><h3 id="m0">TestScript.LoginPage#Testapplication</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
-<h3 id="m1">TestScript.PreAdm_HomePage#Test</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
+<h2>Test</h2><h3 id="m0">TestScript.Application_form#Test</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
+<h3 id="m1">TestScript.LoginPage#Testapplication</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
+<h3 id="m2">TestScript.PreAdm_HomePage#Test</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
 </body>
 </html>
