+++ test-output/index.html
@@ -92,7 +92,7 @@
               <div class="suite-section-content">
                 <ul>
                   <li>
-                    <span class="method-stats">2 methods,   2 passed</span>
+                    <span class="method-stats">3 methods,   3 passed</span>
                   </li>
                   <li>
                     <span class="method-list-title passed">Passed methods</span>
@@ -104,6 +104,11 @@
                       <span>
                         <img src="passed.png" width="3%"/>
                         <a href="#" panel-name="suite-Suite" title="TestScript.PreAdm_HomePage" class="method navigator-link" hash-for-method="Test">Test</a> <!-- method navigator-link -->
+                      </span>
+                      <br/>
+                      <span>
+                        <img src="passed.png" width="3%"/>
+                        <a href="#" panel-name="suite-Suite" title="TestScript.Application_form" class="method navigator-link" hash-for-method="Test">Test</a> <!-- method navigator-link -->
                       </span>
                       <br/>
                       <span>
@@ -153,6 +158,21 @@
               </div> <!-- method -->
             </div> <!-- main-panel-content rounded-window-bottom -->
           </div> <!-- suite-Suite-class-passed -->
+          <div class="suite-Suite-class-passed">
+            <div class="main-panel-header rounded-window-top">
+              <img src="passed.png"/>
+              <span class="class-name">TestScript.Application_form</span>
+            </div> <!-- main-panel-header rounded-window-top -->
+            <div class="main-panel-content rounded-window-bottom">
+              <div class="method">
+                <div class="method-content">
+                  <a name="Test">
+                  </a> <!-- Test -->
+                  <span class="method-name">Test</span>
+                </div> <!-- method-content -->
+              </div> <!-- method -->
+            </div> <!-- main-panel-content rounded-window-bottom -->
+          </div> <!-- suite-Suite-class-passed -->
         </div> <!-- panel Suite -->
         <div panel-name="test-xml-Suite" class="panel">
           <div class="main-panel-header rounded-window-top">
@@ -167,6 +187,7 @@
     &lt;classes&gt;
       &lt;class name=&quot;TestScript.LoginPage&quot;/&gt;
       &lt;class name=&quot;TestScript.PreAdm_HomePage&quot;/&gt;
+      &lt;class name=&quot;TestScript.Application_form&quot;/&gt;
     &lt;/classes&gt;
   &lt;/test&gt; &lt;!-- Test --&gt;
 &lt;/suite&gt; &lt;!-- Suite --&gt;
@@ -180,7 +201,7 @@
           <div class="main-panel-content rounded-window-bottom">
             <ul>
               <li>
-                <span class="test-name">Test (2 classes)</span>
+                <span class="test-name">Test (3 classes)</span>
               </li>
             </ul>
           </div> <!-- main-panel-content rounded-window-bottom -->
@@ -206,20 +227,24 @@ data.addColumn('number', 'Number');
 data.addColumn('string', 'Method');
 data.addColumn('string', 'Class');
 data.addColumn('number', 'Time (ms)');
-data.addRows(2);
+data.addRows(3);
 data.setCell(0, 0, 0)
 data.setCell(0, 1, 'Test')
 data.setCell(0, 2, 'TestScript.PreAdm_HomePage')
-data.setCell(0, 3, 11086);
+data.setCell(0, 3, 10786);
 data.setCell(1, 0, 1)
-data.setCell(1, 1, 'Testapplication')
-data.setCell(1, 2, 'TestScript.LoginPage')
-data.setCell(1, 3, 361);
+data.setCell(1, 1, 'Test')
+data.setCell(1, 2, 'TestScript.Application_form')
+data.setCell(1, 3, 10588);
+data.setCell(2, 0, 2)
+data.setCell(2, 1, 'Testapplication')
+data.setCell(2, 2, 'TestScript.LoginPage')
+data.setCell(2, 3, 380);
 window.suiteTableData['Suite']= { tableData: data, tableDiv: 'times-div-Suite'}
 return data;
 }
               </script>
-              <span class="suite-total-time">Total running time: 11 seconds</span>
+              <span class="suite-total-time">Total running time: 21 seconds</span>
               <div id="times-div-Suite">
               </div> <!-- times-div-Suite -->
             </div> <!-- times-div -->
@@ -234,11 +259,18 @@ return data;
         </div> <!-- panel -->
         <div panel-name="ignored-methods-Suite" class="panel">
           <div class="main-panel-header rounded-window-top">
-            <span class="header-content">1 ignored method</span>
+            <span class="header-content">2 ignored methods</span>
           </div> <!-- main-panel-header rounded-window-top -->
           <div class="main-panel-content rounded-window-bottom">
             <div class="ignored-class-div">
               <span class="ignored-class-name">TestScript.PreAdm_HomePage</span>
+              <div class="ignored-methods-div">
+                <span class="ignored-method-name">Openapplication</span>
+                <br/>
+              </div> <!-- ignored-methods-div -->
+            </div> <!-- ignored-class-div -->
+            <div class="ignored-class-div">
+              <span class="ignored-class-name">TestScript.Application_form</span>
               <div class="ignored-methods-div">
                 <span class="ignored-method-name">Openapplication</span>
                 <br/>
@@ -259,14 +291,21 @@ return data;
               </div> <!-- configuration-test before -->
               <div class="test-method">
                 <span class="method-name">Testapplication</span>
-                <span class="method-start">13653 ms</span>
+                <span class="method-start">11911 ms</span>
               </div> <!-- test-method -->
             </div> <!-- chronological-class -->
             <div class="chronological-class">
               <div class="chronological-class-name">TestScript.PreAdm_HomePage</div> <!-- chronological-class-name -->
               <div class="test-method">
                 <span class="method-name">Test</span>
-                <span class="method-start">14028 ms</span>
+                <span class="method-start">12298 ms</span>
+              </div> <!-- test-method -->
+            </div> <!-- chronological-class -->
+            <div class="chronological-class">
+              <div class="chronological-class-name">TestScript.Application_form</div> <!-- chronological-class-name -->
+              <div class="test-method">
+                <span class="method-name">Test</span>
+                <span class="method-start">23090 ms</span>
               </div> <!-- test-method -->
           </div> <!-- main-panel-content rounded-window-bottom -->
         </div> <!-- panel -->
