+++ test-output/junitreports/TEST-TestScript.Application_form.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!-- Generated by org.testng.reporters.JUnitReportReporter -->
+<testsuite hostname="DESKTOP-9JGPINH" failures="0" tests="1" name="TestScript.Application_form" time="10.588" errors="0" timestamp="2023-09-20T09:39:26 IST" skipped="0">
+  <testcase classname="TestScript.Application_form" name="Test" time="10.588"/>
+</testsuite> <!-- TestScript.Application_form -->
