+++ test-output/junitreports/TEST-TestScript.LoginPage.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!-- Generated by org.testng.reporters.JUnitReportReporter -->
-<testsuite hostname="DESKTOP-9JGPINH" failures="0" tests="1" name="TestScript.LoginPage" time="0.361" errors="0" timestamp="2023-09-14T09:50:36 IST" skipped="0">
-  <testcase classname="TestScript.LoginPage" name="Testapplication" time="0.361"/>
+<testsuite hostname="DESKTOP-9JGPINH" failures="0" tests="1" name="TestScript.LoginPage" time="0.380" errors="0" timestamp="2023-09-20T09:39:26 IST" skipped="0">
+  <testcase classname="TestScript.LoginPage" name="Testapplication" time="0.380"/>
 </testsuite> <!-- TestScript.LoginPage -->
