+++ test-output/old/Suite/classes.html
@@ -26,6 +26,28 @@
 <td align='center' colspan='3'>@AfterClass</td>
 </tr>
 <tr>
+<td>TestScript.Application_form</td>
+<td>&nbsp;</td><td>&nbsp;</td></tr>
+<tr>
+<td align='center' colspan='3'>@Test</td>
+</tr>
+<tr>
+<td>&nbsp;</td>
+<td>Test</td>
+<td>&nbsp;</td></tr>
+<tr>
+<td align='center' colspan='3'>@BeforeClass</td>
+</tr>
+<tr>
+<td align='center' colspan='3'>@BeforeMethod</td>
+</tr>
+<tr>
+<td align='center' colspan='3'>@AfterMethod</td>
+</tr>
+<tr>
+<td align='center' colspan='3'>@AfterClass</td>
+</tr>
+<tr>
 <td>TestScript.LoginPage</td>
 <td>&nbsp;</td><td>&nbsp;</td></tr>
 <tr>
