+++ test-output/old/Suite/methods-not-run.html
@@ -1,3 +1,4 @@
 <h2>Methods that were not run</h2><table>
 <tr><td>Generic.BaseClass.Openapplication</td></tr>
+<tr><td>Generic.BaseClass.Openapplication</td></tr>
 </table>
