+++ test-output/old/Suite/methods.html
@@ -1,10 +1,12 @@
 <h2>Methods run, sorted chronologically</h2><h3>&gt;&gt; means before, &lt;&lt; means after</h3><p/><br/><em>Suite</em><p/><small><i>(Hover the method name to see the test class name)</i></small><p/>
 <table border="1">
 <tr><th>Time</th><th>Delta (ms)</th><th>Suite<br>configuration</th><th>Test<br>configuration</th><th>Class<br>configuration</th><th>Groups<br>configuration</th><th>Method<br>configuration</th><th>Test<br>method</th><th>Thread</th><th>Instances</th></tr>
-<tr bgcolor="ffe6dc">  <td>23/09/14 09:50:24</td>   <td>0</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="LoginPage.Testapplication()[pri:0, instance:TestScript.LoginPage@72e5a8e]">Testapplication</td> 
-  <td>main@1732238286</td>   <td></td> </tr>
-<tr bgcolor="ad6d96">  <td>23/09/14 09:50:10</td>   <td>-13645</td> <td>&nbsp;</td><td title="&gt;&gt;BaseClass.Openapplication()[pri:0, instance:TestScript.LoginPage@72e5a8e]">&gt;&gt;Openapplication</td> 
-<td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td>  <td>main@1732238286</td>   <td></td> </tr>
-<tr bgcolor="f2bbc1">  <td>23/09/14 09:50:24</td>   <td>375</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="PreAdm_HomePage.Test()[pri:0, instance:TestScript.PreAdm_HomePage@771a660]">Test</td> 
-  <td>main@1732238286</td>   <td></td> </tr>
+<tr bgcolor="ffe6dc">  <td>23/09/20 09:39:04</td>   <td>0</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="LoginPage.Testapplication()[pri:0, instance:TestScript.LoginPage@54e1c68b]">Testapplication</td> 
+  <td>main@1107412069</td>   <td></td> </tr>
+<tr bgcolor="f2bbc1">  <td>23/09/20 09:39:05</td>   <td>387</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="PreAdm_HomePage.Test()[pri:0, instance:TestScript.PreAdm_HomePage@2d52216b]">Test</td> 
+  <td>main@1107412069</td>   <td></td> </tr>
+<tr bgcolor="8aa092">  <td>23/09/20 09:38:52</td>   <td>-11903</td> <td>&nbsp;</td><td title="&gt;&gt;BaseClass.Openapplication()[pri:0, instance:TestScript.LoginPage@54e1c68b]">&gt;&gt;Openapplication</td> 
+<td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td>  <td>main@1107412069</td>   <td></td> </tr>
+<tr bgcolor="e965f5">  <td>23/09/20 09:39:15</td>   <td>11179</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="Application_form.Test()[pri:0, instance:TestScript.Application_form@242b836]">Test</td> 
+  <td>main@1107412069</td>   <td></td> </tr>
 </table>
