+++ test-output/old/Suite/toc.html
@@ -9,11 +9,11 @@
 <table border='1' width='100%'>
 <tr valign='top'>
 <td>1 test</td>
-<td><a target='mainFrame' href='classes.html'>2 classes</a></td>
-<td>2 methods:<br/>
+<td><a target='mainFrame' href='classes.html'>3 classes</a></td>
+<td>3 methods:<br/>
 &nbsp;&nbsp;<a target='mainFrame' href='methods.html'>chronological</a><br/>
 &nbsp;&nbsp;<a target='mainFrame' href='methods-alphabetical.html'>alphabetical</a><br/>
-&nbsp;&nbsp;<a target='mainFrame' href='methods-not-run.html'>not run (1)</a></td>
+&nbsp;&nbsp;<a target='mainFrame' href='methods-not-run.html'>not run (2)</a></td>
 </tr>
 <tr>
 <td><a target='mainFrame' href='groups.html'>0 group</a></td>
@@ -22,7 +22,7 @@
 </tr></table>
 <table width='100%' class='test-passed'>
 <tr><td>
-<table style='width: 100%'><tr><td valign='top'>Test (2/0/0)</td><td valign='top' align='right'>
+<table style='width: 100%'><tr><td valign='top'>Test (3/0/0)</td><td valign='top' align='right'>
   <a href='Test.html' target='mainFrame'>Results</a>
 </td></tr></table>
 </td></tr><p/>
