+++ test-output/old/index.html
@@ -4,6 +4,6 @@
 </head><body>
 <h2><p align='center'>Test results</p></h2>
 <table border='1' width='100%' class='main-page'><tr><th>Suite</th><th>Passed</th><th>Failed</th><th>Skipped</th><th>testng.xml</th></tr>
-<tr align='center' class='invocation-passed'><td><em>Total</em></td><td><em>2</em></td><td><em>0</em></td><td><em>0</em></td><td>&nbsp;</td></tr>
+<tr align='center' class='invocation-passed'><td><em>Total</em></td><td><em>3</em></td><td><em>0</em></td><td><em>0</em></td><td>&nbsp;</td></tr>
 <tr align='center' class='invocation-passed'><td><a href='Suite/index.html'>Suite</a></td>
-<td>2</td><td>0</td><td>0</td><td><a href='Suite/testng.xml.html'>Link</a></td></tr></table></body></html>
+<td>3</td><td>0</td><td>0</td><td><a href='Suite/testng.xml.html'>Link</a></td></tr></table></body></html>
