+++ testng.xml
@@ -1,10 +1,11 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd">
-<suite name="Suite" parallel="none">
-  <test  name="Test">
+<suite name="Suite">
+  <test thread-count="5" name="Test">
     <classes>
       <class name="TestScript.LoginPage"/>
-      <class name="TestScript.PreAdm_HomePage"/>  
+      <class name="TestScript.PreAdm_HomePage"/>
+      <class name="TestScript.Application_form"/>
     </classes>
   </test> <!-- Test -->
 </suite> <!-- Suite -->
